[chat-updates-1] added date of message in message

+++ helpers/common.ts
@@ -1,4 +1,4 @@
-function shuffleArray(array:any[]) {
+function shuffleArray(array: any[]) {
   for (let i = array.length - 1; i > 0; i--) {
     const j = Math.floor(Math.random() * (i + 1)); // Generate a random index between 0 and i
     [array[i], array[j]] = [array[j], array[i]]; // Swap elements at i and j
@@ -8,8 +8,8 @@ function shuffleArray(array:any[]) {
 function readableFormatDate(dateString: string) {
   const date = new Date(dateString);
 
-  const options:Intl.DateTimeFormatOptions = {
-    month: "long",
+  const options: Intl.DateTimeFormatOptions = {
+    month: "short",
     day: "numeric",
     hour: "2-digit",
     minute: "2-digit",
@@ -19,4 +19,4 @@ function readableFormatDate(dateString: string) {
   const formattedDate = date.toLocaleString("en-US", options);
   return formattedDate;
 }
-export {  readableFormatDate ,shuffleArray};
+export { readableFormatDate, shuffleArray };
